move smart contract tests
